Move Rx classes under version namespace to prevent collisions
