+++ build.gradle
@@ -3,8 +3,8 @@ apply from: rootProject.file('config/hooks/install-git-hooks.gradle')
 
 buildscript {
   ext.versions = ['minSdk'        : 14,
-                  'compileSdk'    : 25,
-                  'buildTools'    : '26.0.2',
+                  'compileSdk'    : 27,
+                  'buildTools'    : '27.0.3',
 
                   'supportLibrary': '25.3.1',]
 
@@ -27,7 +27,7 @@ buildscript {
     }
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.0.0'
+    classpath 'com.android.tools.build:gradle:3.1.2'
     classpath "gradle.plugin.nl.javadude.gradle.plugins:license-gradle-plugin:0.14.0"
   }
 }
