+++ config/quality_android.gradle
@@ -1,6 +1,12 @@
 apply plugin: 'findbugs'
 apply plugin: 'pmd'
 
+android {
+  lintOptions {
+    disable "GoogleAppIndexingWarning", "AllowBackup"
+  }
+}
+
 task findbugs(type: FindBugs) {
   ignoreFailures = false
   effort = "max"
@@ -40,4 +46,4 @@ task pmd(type: Pmd) {
   }
 }
 
-check.dependsOn 'findbugs', 'pmd'
+check.dependsOn 'assemble', 'findbugs', 'pmd'
