+++ grox-core-rx/src/main/java/com/groupon/grox/rxjava1/StoreOnSubscribe.java
@@ -13,21 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.groupon.grox;
+package com.groupon.grox.rxjava1;
 
-import io.reactivex.ObservableEmitter;
-import io.reactivex.ObservableOnSubscribe;
-import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Cancellable;
-import java.util.concurrent.atomic.AtomicBoolean;
+import com.groupon.grox.Store;
+import rx.Emitter;
+import rx.functions.Action1;
+import rx.functions.Cancellable;
 
 /**
- * Internal subscriber to a store's observable. It basically allows to unsubscribe from the store
- * when the observable is disposed.
+ * Internal subscriber to a store's obersvable. It basically allows to unsubscribe from the store
+ * when the observable is unsubscribed from.
  *
  * @param <STATE> the class of the the state of the store.
  */
-final class StoreOnSubscribe<STATE> implements ObservableOnSubscribe<STATE> {
+final class StoreOnSubscribe<STATE> implements Action1<Emitter<STATE>> {
   private final Store<STATE> store;
 
   StoreOnSubscribe(Store<STATE> store) {
@@ -35,12 +34,14 @@ final class StoreOnSubscribe<STATE> implements ObservableOnSubscribe<STATE> {
   }
 
   @Override
-  public void subscribe(ObservableEmitter<STATE> emitter) throws Exception {
+  public void call(Emitter<STATE> stateEmitter) {
 
     //the internal listener to the store.
     Store.StateChangeListener<STATE> listener =
-        emitter::onNext;
-    emitter.setCancellable(() -> store.unsubscribe(listener));
+        stateEmitter::onNext;
+
+    stateEmitter.setCancellation(() -> store.unsubscribe(listener));
+
     store.subscribe(listener);
   }
 }
