+++ grox-sample/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Grox sample</string>
+    <string name="change_color">Change color</string>
 </resources>
